refactor hello world example block
