+++ example/src/blocks/hello-world/block.json
@@ -1,0 +1,18 @@
+{
+	"name": "example/hello-world",
+	"title": "Hello World",
+    "description": "Example Block to show the Post Picker in usage",
+    "icon": "smiley",
+    "category": "common",
+    "example": {},
+    "supports": {
+        "html": false
+    },
+    "attributes": {
+        "selectedPost": {
+            "type": "array"
+        }
+    },
+    "variations": [],
+	"editorScript": "file:./index.js"
+}
